+++ apps/api/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @express-learn/api
 
+## 2.0.0
+
+### Major Changes
+
+- added authorization and authentication, cli update
+
+### Minor Changes
+
+- added authentication and authorization
+
 ## 1.0.1
 
 ### Patch Changes